Extract startup URL sanitiser module
Move the inline URL sanitisation helpers out of startup::boot into a file-backed startup::url_sanitize module so CodeScene tracks the URL logic separately from the boot-screen module.

Update the boot-screen call sites and test import to use the file-backed module, and refresh the snapshot assertion metadata after the assertion line moved.
