+++ src/startup/mod.rs
@@ -18,6 +18,8 @@ pub(crate) mod phases;
 pub(crate) mod run;
 /// Shared start/run sequencing helpers for the agent loop.
 pub(crate) mod run_flow;
+/// URL sanitisation helpers for safe display of startup URLs.
+pub(crate) mod url_sanitize;
 /// Unix-only runtime management (SIGHUP hot-reload).
 #[cfg(unix)]
 pub(crate) mod unix_runtime;
